requirements.txt
aiodns==3.2.0
async_timeout==4.0.3
brotli==1.1.0
brotlicffi==1.1.0.0
cffi==1.15.1
ConfigParser==6.0.1
contextlib2==21.6.0
cryptography==42.0.5
Cython==0.29.35
defusedxml==0.7.1
disco==1.40.4
dl==0.1.0
docutils==0.20.1
email_validator==2.1.1
fido2==1.1.3
grandalf==0.8
grpc_reflection==1.0.0
grpc_tools==1.0.0
gspread==6.1.0
gssapi==1.8.3
gunicorn==21.2.0
h2==4.1.0
HTMLParser==0.0.2
hypothesis==6.100.1
ipython==8.12.3
ipywidgets==8.0.6
Jinja2==3.1.2
jnius==1.1.0
js==1.0
keyring==25.1.0
langchain_text_splitters==0.0.1
lockfile==0.12.2
matplotlib==3.8.3
mock==5.1.0
mysql_connector_python==8.0.33
numpy==1.24.3
oauth2client==4.1.3
oci==2.125.2
olefile==0.47
openai==1.16.2
pandas==2.2.1
pygraphviz==1.12
pyodide==0.0.2
pyOpenSSL==24.1.0
pyppeteer==2.0.0
PyQt6==6.6.1
PySide6==6.7.0
pytest==8.1.1
python-dotenv==1.0.1
pyu2f==0.1.5
pywin32==306
railroad==0.5.0
scikit_learn==1.4.0
setuptools_scm==8.0.4
simplejson==3.19.2
slack_sdk==3.27.1
Sphinx==7.2.6
sqlparse==0.4.4
sspi==0.1.0
streamlit==1.32.2
streamlit_scrollable_textbox==0.0.3
tensorflow==2.16.1
tensorflow_intel==2.16.1
thread==1.1.1
toml==0.10.2
tornado==6.4
trove_classifiers==2024.3.25
truststore==0.8.0
urllib3_secure_extra==0.1.0
uvloop==0.19.0
wcwidth==0.2.6
xmlrpclib==1.0.1
zstandard==0.22.0
